test: consolidate duplicate test files and remove overlapping tests
Merge sync_ui_utils into sync_ui_logic, summary_logic into summary_edge_cases,
accounts_api into accounts_visibility, and remove duplicate getScrapeRetries
tests from the integration test file.
