+++ app/tests/accounts_visibility.test.ts
@@ -18,7 +18,7 @@ vi.mock('../utils/logger.js', () => ({
     }
 }));
 
-describe('Accounts Visibility API', () => {
+describe('Accounts API', () => {
     let mockClient: any;
     let mockReq: any;
     let mockRes: any;
@@ -43,6 +43,51 @@ describe('Accounts Visibility API', () => {
 
     afterEach(() => {
         vi.restoreAllMocks();
+    });
+
+    describe('GET /api/accounts', () => {
+        it('should return a list of accounts with balances', async () => {
+            mockReq = {
+                method: 'GET',
+                query: {}
+            };
+
+            const mockRows = [
+                {
+                    id: 1,
+                    vendor: 'hapoalim',
+                    account_number: '123456',
+                    balance: 5000.5,
+                    balance_updated_at: '2023-01-01T10:00:00Z',
+                    credential_nickname: 'Main Bank',
+                    credential_vendor: 'hapoalim',
+                    custom_bank_account_nickname: null,
+                    mapped_card_nickname: null,
+                    mapped_card_vendor: null,
+                    credential_id: 10,
+                    linked_bank_account_id: null,
+                    custom_bank_account_number: null
+                }
+            ];
+
+            mockClient.query.mockResolvedValue({
+                rows: mockRows,
+                rowCount: 1
+            });
+
+            await indexHandler(mockReq, mockRes);
+
+            expect(mockRes.status).toHaveBeenCalledWith(200);
+            const response = mockRes.json.mock.calls[0][0];
+            expect(response).toHaveLength(1);
+            expect(response[0]).toMatchObject({
+                id: 1,
+                vendor: 'hapoalim',
+                balance: 5000.5,
+                nickname: 'Main Bank',
+                last4: '3456'
+            });
+        });
     });
 
     describe('PATCH /api/accounts/[id]', () => {
